Move drag context to HOC, improve sorting HOC
