+++ src/pages/questions/create.js
@@ -7,7 +7,7 @@ import { intlShape } from 'react-intl'
 
 import { TeacherLayout } from '../../components/layouts'
 import { QuestionCreationForm } from '../../components/forms'
-import { pageWithIntl, withData } from '../../lib'
+import { pageWithIntl, withData, withDnD } from '../../lib'
 import { CreateQuestionMutation } from '../../graphql/mutations'
 import { QuestionListQuery, TagListQuery } from '../../graphql/queries'
 
@@ -42,6 +42,7 @@ const CreateQuestion = ({
 CreateQuestion.propTypes = propTypes
 
 export default compose(
+  withDnD,
   withData,
   pageWithIntl,
   graphql(TagListQuery),
